First steps in split
