+++ src/datascience-ui/interactive-common/contentPanel.tsx
@@ -1,14 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 'use strict';
-import * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api';
 import * as React from 'react';
 
-import { ErrorBoundary } from '../react-common/errorBoundary';
-import { IKeyboardEvent } from '../react-common/event';
-import { getSettings } from '../react-common/settingsReactSide';
-import { Cell, ICellViewModel } from './cell';
+import { Identifiers } from '../../client/datascience/constants';
 import { InputHistory } from './inputHistory';
+import { ICellViewModel } from './mainState';
 
 // See the discussion here: https://github.com/Microsoft/tslint-microsoft-contrib/issues/676
 // tslint:disable: react-this-binding-issue
@@ -18,31 +15,15 @@ const throttle = require('lodash/throttle') as typeof import('lodash/throttle');
 export interface IContentPanelProps {
     baseTheme: string;
     cellVMs: ICellViewModel[];
+    newCellVM?: ICellViewModel;
     history: InputHistory;
     testMode?: boolean;
     codeTheme: string;
     submittedText: boolean;
     skipNextScroll: boolean;
-    monacoTheme: string | undefined;
-    editorOptions?: monacoEditor.editor.IEditorOptions;
     editable: boolean;
-    editExecutionCount?: string;
-    editorMeasureClassName?: string;
-    newCellVM?: ICellViewModel;
-    selectedCell?: string;
-    focusedCell?: string;
-    allowsMarkdownEditing?: boolean;
-    onCodeChange(changes: monacoEditor.editor.IModelContentChange[], cellId: string, modelId: string): void;
-    onCodeCreated(code: string, file: string, cellId: string, modelId: string): void;
-    openLink(uri: monacoEditor.Uri): void;
-    expandImage(imageHtml: string): void;
-    selectCell?(cellId: string): void;
-    clickCell?(cellId: string): void;
-    doubleClickCell?(cellId: string): void;
-    focusCell?(cellId: string): void;
-    unfocusCell?(cellId: string): void;
-    keyDownCell?(cellId: string, e: IKeyboardEvent): void;
-    renderCellToolbar(cellId: string): JSX.Element[] | null;
+    renderCell(cellVM: ICellViewModel, index: number, containerRef?: React.RefObject<HTMLDivElement>): JSX.Element | null;
+    focusCell(cellVM: ICellViewModel, focusCode: boolean): void;
     onRenderCompleted?(cells: (HTMLDivElement | null)[]): void;
     scrollToBottom(div: HTMLDivElement): void;
 }
@@ -50,7 +31,6 @@ export interface IContentPanelProps {
 export class ContentPanel extends React.Component<IContentPanelProps> {
     private bottomRef: React.RefObject<HTMLDivElement> = React.createRef<HTMLDivElement>();
     private containerRef: React.RefObject<HTMLDivElement> = React.createRef<HTMLDivElement>();
-    private cellRefs: Map<string, React.RefObject<Cell>> = new Map<string, React.RefObject<Cell>>();
     private cellContainerRefs: Map<string, React.RefObject<HTMLDivElement>> = new Map<string, React.RefObject<HTMLDivElement>>();
     private throttledScrollIntoView = throttle(this.scrollIntoView.bind(this), 100);
     constructor(prop: IContentPanelProps) {
@@ -100,77 +80,32 @@ export class ContentPanel extends React.Component<IContentPanelProps> {
 
     public focusCell(cellId: string, focusCode: boolean) {
         const ref = this.cellContainerRefs.get(cellId);
+        const vm = this.props.cellVMs.find(c => c.cell.id === cellId) || (cellId === Identifiers.EditCellId) ? this.props.newCellVM : undefined;
         if (ref && ref.current) {
             ref.current.scrollIntoView({ behavior: 'auto', block: 'nearest', inline: 'nearest' });
-            const cellRef = this.cellRefs.get(cellId);
-            if (cellRef && cellRef.current) {
-                cellRef.current.giveFocus(focusCode);
+            if (vm) {
+                this.props.focusCell(vm, focusCode);
             }
         }
     }
 
     private renderCells = () => {
-        const maxOutputSize = getSettings().maxOutputSize;
-        const maxTextSize = maxOutputSize && maxOutputSize < 10000 && maxOutputSize > 0 ? maxOutputSize : undefined;
-        const baseTheme = getSettings().ignoreVscodeTheme ? 'vscode-light' : this.props.baseTheme;
-
-        return this.props.cellVMs.map((cellVM: ICellViewModel, index: number) =>
-            this.renderCell(cellVM, index, baseTheme, maxTextSize, false));
+        return this.props.cellVMs.map((cellVM: ICellViewModel, index: number) => {
+            let ref: React.RefObject<HTMLDivElement> | undefined;
+            if (!this.cellContainerRefs.has(cellVM.cell.id)) {
+                ref = React.createRef<HTMLDivElement>();
+                this.cellContainerRefs.set(cellVM.cell.id, ref);
+            }
+            this.props.renderCell(cellVM, index, ref!);
+        });
     }
 
     private renderEdit = () => {
         if (this.props.editable && this.props.newCellVM) {
-            const maxOutputSize = getSettings().maxOutputSize;
-            const maxTextSize = maxOutputSize && maxOutputSize < 10000 && maxOutputSize > 0 ? maxOutputSize : undefined;
-            const baseTheme = getSettings().ignoreVscodeTheme ? 'vscode-light' : this.props.baseTheme;
-            return this.renderCell(this.props.newCellVM, 0, baseTheme, maxTextSize, true);
+            return this.props.renderCell(this.props.newCellVM, 0, undefined);
         } else {
             return null;
         }
-    }
-
-    private renderCell(cellVM: ICellViewModel, index: number, baseTheme: string, maxTextSize: number | undefined, showWatermark: boolean): JSX.Element {
-        const cellRef = React.createRef<Cell>();
-        const ref = React.createRef<HTMLDivElement>();
-        this.cellRefs.set(cellVM.cell.id, cellRef);
-        this.cellContainerRefs.set(cellVM.cell.id, ref);
-        return (
-            <div key={index} id={cellVM.cell.id} ref={ref}>
-                <ErrorBoundary key={index}>
-                    <Cell
-                        ref={cellRef}
-                        role='listitem'
-                        editorOptions={this.props.editorOptions}
-                        history={undefined}
-                        maxTextSize={maxTextSize}
-                        autoFocus={false}
-                        testMode={this.props.testMode}
-                        cellVM={cellVM}
-                        baseTheme={baseTheme}
-                        codeTheme={this.props.codeTheme}
-                        allowCollapse={!this.props.editable}
-                        showWatermark={showWatermark}
-                        editExecutionCount={this.props.editExecutionCount}
-                        onCodeChange={this.props.onCodeChange}
-                        onCodeCreated={this.props.onCodeCreated}
-                        monacoTheme={this.props.monacoTheme}
-                        openLink={this.props.openLink}
-                        expandImage={this.props.expandImage}
-                        editorMeasureClassName={this.props.editorMeasureClassName}
-                        selectedCell={this.props.selectedCell}
-                        focusedCell={this.props.focusedCell}
-                        onClick={this.props.clickCell}
-                        focused={this.props.focusCell}
-                        unfocused={this.props.unfocusCell}
-                        keyDown={this.props.keyDownCell}
-                        allowsMarkdownEditing={this.props.allowsMarkdownEditing}
-                        onDoubleClick={this.props.doubleClickCell}
-                        renderCellToolbar={this.props.renderCellToolbar}
-                        showLineNumbers={cellVM.showLineNumbers}
-                        hideOutput={cellVM.hideOutput}
-                    />
-                </ErrorBoundary>
-            </div>);
     }
 
     private scrollIntoView() {
