+++ src/test/datascience/notebook.functional.test.ts
@@ -41,8 +41,7 @@ import {
     IKnownSearchPathsForInterpreters,
     PythonInterpreter
 } from '../../client/interpreter/contracts';
-import { ICellViewModel } from '../../datascience-ui/interactive-common/cell';
-import { generateTestState } from '../../datascience-ui/interactive-common/mainState';
+import { generateTestState, ICellViewModel } from '../../datascience-ui/interactive-common/mainState';
 import { asyncDump } from '../common/asyncDump';
 import { sleep } from '../core';
 import { DataScienceIocContainer } from './dataScienceIocContainer';
